Handle invite acceptance statuses: - Redirect to sign-in on 404 - Render “Not authorized” view on 401 - Redirect to auth callback on 200
